+++ fabricpc/training/__init__.py
@@ -18,7 +18,9 @@ from fabricpc.training.trainer import (
     convert_batch,
     create_causal_mask,
     evaluate,
+    grad_denominator,
     make_train_step,
+    pc_weight_gradients,
     train,
 )
 
@@ -30,6 +32,8 @@ __all__ = [
     "build_clamps",
     "convert_batch",
     "create_causal_mask",
+    "grad_denominator",
+    "pc_weight_gradients",
     "TrainResult",
     "EpochContext",
     "EvalMetric",